Add cross-platform build support and freeglut docs, using Cursor

+++ langs/cpp/GameView.h
@@ -20,7 +20,7 @@ class GameView : public GameObserver
 		int viewX2;
 		int viewY2;
 		int showInfo;
-		int milisec;
+		unsigned long milisec;
 		int frameCnt;
 		int copyBoardIndex;
 		int frameRateDelay;
